Renamed Data package to model.
